gotit/requirements.txt
slack-sdk>=3.21.3
notion-client>=2.2.1
python-dotenv>=1.0.0
google-generativeai>=0.3.2
PyYAML>=6.0.1
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2 